add HomeActivity.kt
