+++ DDUBUCK/app/src/main/res/layout/activity_home.xml
@@ -6,12 +6,12 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/teal_200"
-    tools:context=".home.CourseSelectSheetActivity">
+    tools:context=".home.HomeActivity">
 
     <include layout="@layout/map_fragment"
-        android:id="@+id/test_map"/>
+        android:id="@+id/home_map_fragment"/>
     <include layout="@layout/bottom_sheet"
-        android:id="@+id/bottom_sheet"
+        android:id="@+id/home_bottom_sheet"
         android:background="@android:color/white" />
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
